Set up node-jscs via Grunt and pass it
JSCS is a code style checker (as opposed to jshint, which is for
code quality). There are a few small style-related things that
JSHint did (quoatmark, onevar, noempty etc.) but those are being
deprecated in JSHint v3, people should use more sophisticated
tools like node-jscs for this instead. As such this commit
removes moves those options from the jshint configuration.